# docs/swagger.yml
swagger: '2.0'
info:
  description: Initial Setup for Swagger, better description coming in future versions
  version: 1.0.0
  title: Everest API
  termsOfService: URL here
  contact:
    email: tracy@expertel.ca
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: localhost
basePath: /v1
...
schemes:
- https
- http
paths:
...
  '/survey/{id}/answers':
    get:
      tags:
      - answers
      summary: Get a list of all answers for a given survey for this subject
      operationId: getAnswers
      consumes:
        - application/x-www-form-urlencoded
        - text/plain; charset=utf-8
      produces:
        - application/json
      parameters:
        - in: header
          name: X-Everest-Auth
          type: string
          description: JWT that contains data about the current subject
        - in: path
          name: id
          description: id of the survey that is changing
          type: integer
          required: true
      responses:
        '200':
          description: List of answers objects
    post:
      tags:
      - answers
      summary: Add a new set of answers for this survey
      operationId: addAnswers
      consumes:
        - application/x-www-form-urlencoded
        - text/plain; charset=utf-8
      produces:
        - application/json
      parameters:
        - in: header
          name: X-Everest-Auth
          type: string
          description: JWT that contains data about the current subject
        - in: path
          name: id
          description: id of the survey that is changing
          type: integer
          required: true
      responses:
        '200':
          description: Confirmation that answers have been added
...
  /events:
    get:
      tags:
      - event
      summary: Get a list of all Events for a given subject
      operationId: getEvents
      consumes:
        - application/x-www-form-urlencoded
        - text/plain; charset=utf-8
      produces:
        - application/json
      parameters:
        - in: header
          name: X-Everest-Auth
          type: string
          description: JWT that contains data about the current subject
      responses:
        '200':
          description: List of Events objects
          schema:
            $ref: '#/definitions/Event'
...
definitions:
  Meta:
    properties:
      currentTime:
        type: string
  Event: 
    properties:
      meta:
        $ref: '#/definitions/Meta'
      payload:
        type: object
        properties:
          id:
            type: integer
          userId:
            type: string
          eventDate:
            type: string
          eventType:
            type: string
          created:
            type: string
          modified:
            type: string
          canceled:
            type: integer
...